requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
dj-database-url==2.1.0
Django==5.0
django-crispy-forms==2.1
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
